fix: resolve Expo Router and Jest configuration issues
- Move dungeon game files from app directory to components directory

- Create proper route file that imports component

- Prevent test files from being treated as routes by Expo Router

- Maintain proper separation between app routes and component logic

Related to dungeon game feature implementation
